+++ desktop/src/features/settings/ui/SettingsPanels.tsx
@@ -1,4 +1,4 @@
-import { useState, useMemo, useRef } from "react";
+import { useMemo } from "react";
 import {
   BellRing,
   Bot,
@@ -9,14 +9,10 @@ import {
   Keyboard,
   LayoutTemplate,
   LockKeyhole,
-  Monitor,
   MonitorCog,
-  Moon,
-  Search,
   Smartphone,
   Smile,
   Stethoscope,
-  Sun,
   UserRound,
   type LucideIcon,
 } from "lucide-react";
@@ -33,8 +29,21 @@ import {
   NEUTRAL_ACCENT,
   useTheme,
 } from "@/shared/theme/ThemeProvider";
-import { SYNTAX_THEMES, isLightTheme } from "@/shared/theme/theme-loader";
-import { Switch } from "@/shared/ui/switch";
+import {
+  LIGHT_THEMES,
+  SYNTAX_THEMES,
+  type SyntaxThemeName,
+  getThemePair,
+} from "@/shared/theme/theme-loader";
+import {
+  ThemePreviewFrame,
+  type ThemePreviewVars,
+} from "@/shared/theme/ThemePreviewFrame";
+import {
+  getThemeFallbackPreviewVars,
+  useThemePreviewVars,
+  withAccentPreviewVars,
+} from "@/shared/theme/useThemePreviewVars";
 import { ChannelTemplatesSettingsCard } from "./ChannelTemplatesSettingsCard";
 import { DoctorSettingsPanel } from "./DoctorSettingsPanel";
 import { ExperimentalFeaturesCard } from "./ExperimentalFeaturesCard";
@@ -189,37 +198,171 @@ function formatThemeLabel(name: string): string {
     .join(" ");
 }
 
+/**
+ * Categorize themes into three groups:
+ * 1. Paired — themes with both a light and dark variant (auto-switches with system)
+ * 2. Light-only — light themes with no dark counterpart
+ * 3. Dark-only — dark themes with no light counterpart
+ *
+ * For paired themes, we deduplicate by only keeping the light member
+ * (the dark member is shown alongside it as a preview).
+ */
+function useThemeCategories() {
+  return useMemo(() => {
+    const pairedLight: SyntaxThemeName[] = [];
+    const lightOnly: SyntaxThemeName[] = [];
+    const darkOnly: SyntaxThemeName[] = [];
+
+    // Track which themes are the "dark side" of a pair so we skip them
+    const darkPairMembers = new Set<string>();
+    for (const name of SYNTAX_THEMES) {
+      if (LIGHT_THEMES.has(name)) {
+        const pair = getThemePair(name);
+        if (pair) {
+          darkPairMembers.add(pair);
+        }
+      }
+    }
+
+    for (const name of SYNTAX_THEMES) {
+      // Skip dark members of pairs — they'll be shown alongside their light counterpart
+      if (darkPairMembers.has(name)) continue;
+
+      if (LIGHT_THEMES.has(name)) {
+        const pair = getThemePair(name);
+        if (pair) {
+          pairedLight.push(name);
+        } else {
+          lightOnly.push(name);
+        }
+      } else {
+        darkOnly.push(name);
+      }
+    }
+
+    return { pairedLight, lightOnly, darkOnly };
+  }, []);
+}
+
+function PairedThemeTile({
+  isActive,
+  lightName,
+  lightVars,
+  darkVars,
+  onSelect,
+}: {
+  isActive: boolean;
+  lightName: SyntaxThemeName;
+  lightVars: ThemePreviewVars | null;
+  darkVars: ThemePreviewVars | null;
+  onSelect: () => void;
+}) {
+  return (
+    <button
+      aria-pressed={isActive}
+      className={cn(
+        "group flex min-w-0 flex-col rounded-lg border bg-background/70 p-2 text-left transition-colors focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-ring",
+        isActive
+          ? "border-primary text-foreground shadow-sm"
+          : "border-border/70 text-muted-foreground hover:border-border hover:bg-accent/70 hover:text-accent-foreground",
+      )}
+      data-testid={`theme-pair-${lightName}`}
+      onClick={onSelect}
+      type="button"
+    >
+      <div className="flex gap-2">
+        <ThemePreviewFrame className="h-20 w-[120px]" vars={lightVars} />
+        <ThemePreviewFrame className="h-20 w-[120px]" vars={darkVars} />
+      </div>
+
+      <div className="mt-2 flex min-h-6 items-center gap-2 px-1">
+        <span className="min-w-0 flex-1 truncate text-sm font-medium">
+          {formatThemeLabel(
+            lightName.replace(/-(?:light|latte|dawn|lotus|ochin)$/, ""),
+          )}
+        </span>
+        {isActive ? <Check className="h-4 w-4 shrink-0 text-primary" /> : null}
+      </div>
+    </button>
+  );
+}
+
+function SingleThemeTile({
+  isActive,
+  name,
+  vars,
+  onSelect,
+}: {
+  isActive: boolean;
+  name: SyntaxThemeName;
+  vars: ThemePreviewVars | null;
+  onSelect: () => void;
+}) {
+  return (
+    <button
+      aria-pressed={isActive}
+      className={cn(
+        "group flex w-[174px] min-w-0 flex-col rounded-lg border bg-background/70 p-2 text-left transition-colors focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-ring",
+        isActive
+          ? "border-primary text-foreground shadow-sm"
+          : "border-border/70 text-muted-foreground hover:border-border hover:bg-accent/70 hover:text-accent-foreground",
+      )}
+      data-testid={`theme-option-${name}`}
+      onClick={onSelect}
+      type="button"
+    >
+      <ThemePreviewFrame vars={vars} />
+
+      <div className="mt-2 flex min-h-6 items-center gap-2 px-1">
+        <span className="min-w-0 flex-1 truncate text-sm font-medium">
+          {formatThemeLabel(name)}
+        </span>
+        {isActive ? <Check className="h-4 w-4 shrink-0 text-primary" /> : null}
+      </div>
+    </button>
+  );
+}
+
 function ThemeSettingsCard() {
   const {
     setTheme,
-    themeName,
     selectedThemeName,
     isDark,
     accentColor,
     setAccentColor,
-    followSystem,
     setFollowSystem,
   } = useTheme();
-  const [search, setSearch] = useState("");
-  const didScrollRef = useRef(false);
-  const activeRef = (node: HTMLButtonElement | null) => {
-    if (node && !didScrollRef.current) {
-      didScrollRef.current = true;
-      node.scrollIntoView({ block: "center" });
-    }
+
+  const previewVarsByTheme = useThemePreviewVars();
+  const { pairedLight, lightOnly, darkOnly } = useThemeCategories();
+
+  const getVars = (name: SyntaxThemeName) =>
+    withAccentPreviewVars(
+      previewVarsByTheme[name] ?? getThemeFallbackPreviewVars(name),
+      accentColor,
+    );
+
+  /** Check if a paired theme (by its light member) is the active selection */
+  const isPairActive = (lightName: SyntaxThemeName) => {
+    const darkName = getThemePair(lightName);
+    return selectedThemeName === lightName || selectedThemeName === darkName;
   };
 
-  const selectedTheme = selectedThemeName;
+  const handleSelectPair = (lightName: SyntaxThemeName) => {
+    // Selecting a paired theme auto-enables follow-system
+    setTheme(lightName);
+    setFollowSystem(true);
+  };
 
-  const filtered = useMemo(() => {
-    const q = search.toLowerCase().trim();
-    if (!q) return SYNTAX_THEMES;
-    return SYNTAX_THEMES.filter((name) => name.includes(q));
-  }, [search]);
+  const handleSelectSingle = (name: SyntaxThemeName) => {
+    setTheme(name);
+    // Single themes don't follow system — they're always one mode
+    setFollowSystem(false);
+  };
 
   return (
     <section
-      className="flex min-h-0 flex-1 flex-col"
+      className="flex min-h-0 flex-1 flex-col overflow-y-auto"
       data-testid="settings-theme"
     >
       <SettingsSectionHeader
@@ -227,115 +370,108 @@ function ThemeSettingsCard() {
         description="Choose a theme for Buzz."
       />
 
-      <div className="relative mb-3 shrink-0">
-        <Search className="pointer-events-none absolute left-3 top-1/2 h-4 w-4 -translate-y-1/2 text-muted-foreground" />
-        <input
-          autoCapitalize="none"
-          autoCorrect="off"
-          className="w-full rounded-lg border border-border/70 bg-background/70 py-2 pl-9 pr-3 text-sm placeholder:text-muted-foreground focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-ring"
-          onChange={(e) => setSearch(e.target.value)}
-          placeholder="Search themes..."
-          spellCheck={false}
-          type="text"
-          value={search}
-        />
+      {/* Section 1: Paired themes (follow system) */}
+      <div className="mb-6">
+        <h3 className="mb-1 text-sm font-medium text-foreground">
+          Adapts to system
+        </h3>
+        <p className="mb-3 text-2xs text-muted-foreground">
+          Automatically switches between light and dark with your system
+          preferences.
+        </p>
+        <div className="grid grid-cols-[repeat(auto-fill,minmax(260px,1fr))] gap-3">
+          {pairedLight.map((lightName) => {
+            const darkName = getThemePair(lightName);
+            if (!darkName) return null;
+            return (
+              <PairedThemeTile
+                darkVars={getVars(darkName)}
+                isActive={isPairActive(lightName)}
+                key={lightName}
+                lightName={lightName}
+                lightVars={getVars(lightName)}
+                onSelect={() => handleSelectPair(lightName)}
+              />
+            );
+          })}
+        </div>
       </div>
 
-      <div className="min-h-0 flex-1 overflow-y-auto rounded-lg border border-border/70 bg-background/70">
-        {filtered.length === 0 ? (
-          <p className="px-3 py-4 text-center text-sm text-muted-foreground">
-            No themes match your search.
+      {/* Section 2: Light-only themes */}
+      {lightOnly.length > 0 && (
+        <div className="mb-6">
+          <h3 className="mb-1 text-sm font-medium text-foreground">Light</h3>
+          <p className="mb-3 text-2xs text-muted-foreground">
+            Always uses a light appearance.
           </p>
-        ) : (
-          filtered.map((name) => {
-            const isActive = selectedTheme === name;
-            const isEffective = themeName === name;
-            const light = isLightTheme(name);
+          <div className="grid grid-cols-[repeat(auto-fill,174px)] gap-3">
+            {lightOnly.map((name) => (
+              <SingleThemeTile
+                isActive={selectedThemeName === name}
+                key={name}
+                name={name}
+                onSelect={() => handleSelectSingle(name)}
+                vars={getVars(name)}
+              />
+            ))}
+          </div>
+        </div>
+      )}
+
+      {/* Section 3: Dark-only themes */}
+      {darkOnly.length > 0 && (
+        <div className="mb-6">
+          <h3 className="mb-1 text-sm font-medium text-foreground">Dark</h3>
+          <p className="mb-3 text-2xs text-muted-foreground">
+            Always uses a dark appearance.
+          </p>
+          <div className="grid grid-cols-[repeat(auto-fill,174px)] gap-3">
+            {darkOnly.map((name) => (
+              <SingleThemeTile
+                isActive={selectedThemeName === name}
+                key={name}
+                name={name}
+                onSelect={() => handleSelectSingle(name)}
+                vars={getVars(name)}
+              />
+            ))}
+          </div>
+        </div>
+      )}
+
+      {/* Accent color picker */}
+      <div className="mt-2 shrink-0 pb-2">
+        <h3 className="mb-2 text-sm font-medium">Accent color</h3>
+        <div className="flex flex-wrap gap-2">
+          {ACCENT_COLORS.map((color) => {
+            const isNeutral = color.value === NEUTRAL_ACCENT;
+            const swatchColor = isNeutral
+              ? "hsl(var(--foreground))"
+              : color.value;
+            const checkClassName =
+              isNeutral && isDark ? "text-black" : "text-white";
 
             return (
               <button
-                aria-pressed={isActive}
                 className={cn(
-                  "flex w-full items-center gap-3 px-3 py-2 text-left text-sm transition-colors focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-inset focus-visible:ring-ring",
-                  isActive
-                    ? "bg-primary/10 text-foreground"
-                    : isEffective && followSystem
-                      ? "bg-primary/5 text-foreground"
-                      : "text-muted-foreground hover:bg-accent hover:text-accent-foreground",
+                  "flex h-7 w-7 items-center justify-center rounded-full border border-border/50 transition-transform hover:scale-110",
+                  accentColor === color.value &&
+                    "ring-2 ring-ring ring-offset-2 ring-offset-background",
                 )}
-                data-testid={`theme-option-${name}`}
-                key={name}
-                onClick={() => setTheme(name)}
-                ref={isActive ? activeRef : undefined}
+                data-testid={`accent-color-${color.name.toLowerCase()}`}
+                key={color.value}
+                onClick={() => setAccentColor(color.value)}
+                style={{ backgroundColor: swatchColor }}
+                title={color.name}
                 type="button"
               >
-                {light ? (
-                  <Sun className="h-4 w-4 shrink-0" />
-                ) : (
-                  <Moon className="h-4 w-4 shrink-0" />
-                )}
-                <span className="flex-1 truncate">
-                  {formatThemeLabel(name)}
-                </span>
-                {isActive && (
-                  <Check className="h-4 w-4 shrink-0 text-primary" />
-                )}
-                {!isActive && isEffective && followSystem && (
-                  <Monitor className="h-3.5 w-3.5 shrink-0 text-muted-foreground" />
+                {accentColor === color.value && (
+                  <Check className={cn("h-4 w-4", checkClassName)} />
                 )}
               </button>
             );
-          })
-        )}
-      </div>
-
-      <div className="mt-4 flex shrink-0 flex-col gap-3 pb-2 sm:flex-row sm:items-start sm:justify-between">
-        <div className="min-w-0">
-          <h3 className="mb-2 text-sm font-medium">Accent color</h3>
-          <div className="flex flex-wrap gap-2">
-            {ACCENT_COLORS.map((color) => {
-              const isNeutral = color.value === NEUTRAL_ACCENT;
-              const swatchColor = isNeutral
-                ? "hsl(var(--foreground))"
-                : color.value;
-              const checkClassName =
-                isNeutral && isDark ? "text-black" : "text-white";
-
-              return (
-                <button
-                  className={cn(
-                    "flex h-7 w-7 items-center justify-center rounded-full border border-border/50 transition-transform hover:scale-110",
-                    accentColor === color.value &&
-                      "ring-2 ring-ring ring-offset-2 ring-offset-background",
-                  )}
-                  data-testid={`accent-color-${color.name.toLowerCase()}`}
-                  key={color.value}
-                  onClick={() => setAccentColor(color.value)}
-                  style={{ backgroundColor: swatchColor }}
-                  title={color.name}
-                  type="button"
-                >
-                  {accentColor === color.value && (
-                    <Check className={cn("h-4 w-4", checkClassName)} />
-                  )}
-                </button>
-              );
-            })}
-          </div>
+          })}
         </div>
-
-        <label
-          className="flex cursor-pointer items-center gap-3 text-sm font-medium text-foreground"
-          htmlFor="follow-system-switch"
-        >
-          <span className="min-w-0 truncate">Use system setting</span>
-          <Switch
-            checked={followSystem}
-            data-testid="follow-system-toggle"
-            id="follow-system-switch"
-            onCheckedChange={setFollowSystem}
-          />
-        </label>
       </div>
     </section>
   );
